Extracting CategoryList from App component
