+++ react-app/src/App.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react'
+import CategoryList from './components/CategoryList'
 import courses from './data/courses'
 
 const App = () => {
@@ -26,50 +27,11 @@ const App = () => {
     )
   })
 
-  const pluck = (array, key) => {
-    let pluckedArray = []
-    array.map(item => {
-      if (item.hasOwnProperty(key)) {
-        pluckedArray.push(item[key])
-      }
-    })
-    return pluckedArray
-  }
-
-  let uniqueCategories = pluck(courses, 'category').filter((value, index, array) => array.indexOf(value) === index)
-
-  const hideCourses = (courseDetails, category) => {
-    let filteredResult = courseDetails.filter(course => {
-      return course.props.data_cat === category
-    })
-    setFilteredResult(filteredResult)
-  }
-
-  let categoryList = uniqueCategories.map(category => {
-    return (
-      <li
-        className='bold capitalized'
-        key={category}
-        data-cy={category}
-        onClick={() => hideCourses(courseDetails, category)}
-      >{category}
-      </li>
-    )
-  })
-
   return (
     <>
       <div data-cy='app'>
         <h1 className='center'>Course list</h1>
-        <ul data-cy='category-list' className='center'>
-          <li
-            data-cy='all'
-            onClick={() => setFilteredResult()}
-            className='bold capitalized'
-          >All
-          </li>
-          {categoryList}
-        </ul>
+        <CategoryList courseDetails={courseDetails} setFilteredResult={setFilteredResult} />
         {filteredResult ?
           filteredResult
           :
